fix(build): include dev-novel build as part of the global build command
The purpose of this change is to better detect build failures on dev-novel
Right now, dev-novel is only build in the netlify env, which empirically tends to swallow errors.

for example in this build: https://app.netlify.com/sites/angular-instantsearch/deploys/5bd9987f02ed833a3299b33c
```
1:02:57 PM: One CLI for webpack must be installed. These are recommended choices, delivered as separate packages:
1:02:57 PM:  - webpack-cli (https://github.com/webpack/webpack-cli)
1:02:57 PM:    The original webpack full-featured CLI.
1:02:57 PM: We will use "yarn" to install the CLI via "yarn add -D".
1:02:57 PM: Do you want to install 'webpack-cli' (yes/no): Done in 316.25s.
```
you can see that the unaswered prompt did not result in an error.
Ideally this should timeout and eventually fail.