fix(RefinementList): introduce showMore prop to hide button
priorly the user could control the display of the show more button with
by acting on the showMoreLimit prop. This internally changed the
canToggleShowMore state which eventually controlled the visibility of the button.

In other flavours it is controlled by an external prop showMore : Boolean.

This PR intends to make the behaviour consitent with other flavours