fix(RefinementList): introduce showMore prop to hide button
priorly the user could control the display of the show more button with
by acting on the showMoreLimit prop. This internally changed the
canToggleShowMore state which eventually controlled the visibility of the button.

In other flavours it is controlled by an external prop showMore : Boolean.

This PR intends to make the behaviour consitent with other flavours

Squashed commit of the following:

commit f836465b4cef530f2dc044244a9cf513dfebf436
Author: Youcef Mammar <youcef.mammar@algolia.com>
Date:   Mon Jan 14 12:41:53 2019 +0100

    fix(RefinementList): introduce showMore prop to hide button

    priorly the user could control the display of the show more button with
    by acting on the showMoreLimit prop. This internally changed the
    canToggleShowMore state which eventually controlled the visibility of the button.

    In other flavours it is controlled by an external prop showMore : Boolean.

    This PR intends to make the behaviour consitent with other flavours