chore(examples): migrate media example to v3 (#485)
This is rather fixing the version number in the package.json.
The code of the example has already been migrated.

It didn't look broken because the examples are always built with the "current" angular-instantsearch codebase regardless of the version that is referenced in the package.json. (cf. https://github.com/algolia/angular-instantsearch/blob/756862922ee240aced3d82fa985f4a20dc4b1815/scripts/build-example.sh)