fix(Hits): add test for template support
This PR adds tests for make sure `ais-hits` works with `<ng-template>`
and that it's passing `hits` and `results` like we want to.

This is a fairly common use case and yet we do not test it.
I noticed this as I was attempting to write tests for the insights
integration.