fix(InfiniteHits): improve typing
**Summary**

introduced InfiniteHitsState and InfiniteHitsItem
I could have copy pasted some of the types we have in InstantSearch V3
but they are a believe a little complex and will certainly be subject to
changes.

I'm afraid if we systematically copy paste the types we have currently
in IS V3 we'll just end up maintaining them in two different places or
with a lot of inconsistencies.

Ideally IS V3 would allow importing the types and use them here
directly. In the mean time, the idea is to keep Angular IS types very simple
so they'll be easier to migrate to to IS's complex types.

**Result**
N/A