fix(types): update Hit type
**Summary**
This type spec is copy pasted from the one we currently have in InstantSearch.

https://github.com/algolia/instantsearch.js/blob/65911504b9ae0e2596c75f0ae74133985dc79850/src/types/instantsearch.ts#L20-L71

Generally copy pasting is not a good idea but [this comment](https://github.com/algolia/angular-instantsearch/pull/553/files#r284135258) makes the point that since it's mostely Search API spec, it may not change much.

If this is merged, both of this PR need to be updated
- fix(Hits): use connector transformItems -> https://github.com/algolia/angular-instantsearch/pull/551
- fix(InfiniteHits): improve typing -> https://github.com/algolia/angular-instantsearch/pull/553

**Results**
N/A