feat(core): add support for algoliasearch v4
Until now, only algoliasearch v3 was supported (for server-side rendering). Now both versions are supported by means of a custom request function for either.

I also changed the existing test to be either for v3 or for v4, as well as adding a new test to verify that the custom request works with Angular semantics & gets cachd correctly

the ssr example has been updated to v4, but still works with v3 too (verified locally)