fix(types): add type inference to base-widget
This change will allow type inference on the `state` exposed by the
connector.

This is helpful in templates

```ts
@Component({
  selector: 'app-rating-menu',
  template: `<div> {{state.createURL(1234)}} </div>`
                                     ^^^^
                                     // rgument type 1234 is not assignable to parameter type string
})
export class RatingMenu extends BaseWidget<typeof connectRatingMenu> {...}
```

This is also useful when customizing updateState

```ts
@Component({
  selector: 'app-rating-menu',
})
export class RatingMenu extends BaseWidget<typeof connectRatingMenu> {

	...
	public updateState(state /* type inferred */, isFirstRendering /* type inferred */) => {
		...
	}
}
```

This would be a BREAKING CHANGE but we can decide not make it one, if we
do this only on a TypedBaseWidget then migrate our widget there.